style fix

+++ project/include/timestamps.h
@@ -9,6 +9,6 @@
 typedef struct timespec timespec;
 
 timespec diff(timespec *start, timespec *end);
-uint64_t get_elapsed_time_from_start_us(timespec startTime);
+size_t get_elapsed_time_from_start_us(timespec startTime);
 
 #endif  // PROJECT_INCLUDE_TIMESTAMPS_H_
